assignment-08-DA24M011-Nandhakishore/requirements.txt
mlflow
tensorflow-macos
numpy
pillow
scikit-learn
editdistance
scikit-learn
python-Levenshtein